Move OffsetInformationTest to ui-tooling-data
OffsetInformationTest depends only on
ui-tooling-data so it belongs there.

bug: 170895806
Test: moved tests pass
